creating a new app for thesis assignments
