+++ server/managementSystem/managementSystem/settings.py
@@ -45,6 +45,7 @@ INSTALLED_APPS = [
     'corsheaders',
 
     'teachingAssignment',
+    'thesisAssignment',
 
 ]
