+++ server/managementSystem/managementSystem/urls.py
@@ -9,5 +9,6 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     path('api/token/', MyTokenObtainPairView.as_view(), name='token_obtain_pair'),
     path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
-    path('', include('teachingAssignment.urls'))
+    path('', include('teachingAssignment.urls')),
+    path('thesis/', include('thesisAssignment.urls'))
 ]
